refactor(app): move alot of the sceen logic into a builder and hook
